refactor(move-traits-derive): split into modules
